feat: return 404 when delete todo with invalid id

+++ src/main/java/oocl/example/controller/TodoController.java
@@ -47,6 +47,9 @@ public class TodoController {
 
   @DeleteMapping("/{id}")
   public CommonResult<String> deleteById(@PathVariable int id) {
+    if(todoService.getTodoById(id) == null) {
+      return CommonResult.notFound("Todo not found");
+    }
     todoService.deleteById(id);
     return CommonResult.deleteSuccess();
   }
